'export' is not optional here; add react-jsx.d.ts
